Transcribe Ollie's original SO API gist
See: https://gist.github.com/olivergierke/1583871453d7f2b59b16

Note: SpringStackOverflowClientTests FAILS with a JsonParseException
caused by an HttpMessageNotReadableException. This seems to be due
to binary data coming across the wire.

In any case, it doesn't much matter. We need to re-work this code anyway
so that it fetches data based on our search query, which will look
something like this:

https://api.stackexchange.com/2.2/search?page=1&pagesize=10&order=desc&sort=activity&tagged=spring-mvc%3Bspring-data%3Bspring-security&site=stackoverflow