Generate toString methods using lombok's @ToString
This will not break compilation within the IDE in any case. However, in
order to actually generate the toString method see the correct output
when working within an IDE, the Lombok plugin is required.

The toString method will always be generated and will always work
properly when the application is being run outside of the IDE (i.e.
via `java -jar` or `gradle :sagan-site:bootRun`).

In IDEA, be sure to check "Enable Annotation Processing" in
Compiler > Annotation Processors.

Note that the plan as of this commit is to allow use of lombok
generation only for toString, equals and hashcode methods, because
doing so is very low impact, i.e. does not create any barrier to entry
for folks unfamiliar with Lombok. They do not need to install any
plugins just to get things working in their IDE, etc. Perhaps we'll use
Lombok more extensively in the future, but for now, it's intentionally
limited.

See #149